requirements.txt
streamlit>=1.35.0
PyPDF2>=3.0.1
openai>=1.50.0
pandas>=2.2.0
plotly>=5.17.0
numpy>=1.26.0
scikit-learn>=1.5.0
python-dotenv>=1.0.0